# Performance improvements, high accuracy simulation button, .webp instead of .png, update d3js, update version of r6operator package