--- a/gulpfile.js/gulp/scripts.js
+++ b/gulpfile.js/gulp/scripts.js
@@ -19,6 +19,7 @@ var _ = require('lodash'),
 
 function scripts_start() {
     return gulp.src([
+            conf.paths.src + '/scripts/app.js',
             conf.paths.src + '/scripts/d3.min.js',
             conf.paths.src + '/scripts/neo4jd3.js',
             conf.paths.src + '/scripts/imageLocation.js',