--- a/src/main/scripts/neo4jd3.js
+++ b/src/main/scripts/neo4jd3.js
@@ -736,9 +736,13 @@ function Neo4jD3(_selector, _options) {
                 return 'translate(' + d.source.x + ', ' + d.source.y + ') rotate(' + angle + ')';
             });
 
+            const startTime = performance.now();
             tickRelationshipsTexts();
             tickRelationshipsOutlines();
             tickRelationshipsOverlays();
+
+            const duration = performance.now() - startTime;
+            console.log(`tick took ${duration}ms`);
         }
     }