--- a/src/main/scripts/neo4jd3.js
+++ b/src/main/scripts/neo4jd3.js
@@ -698,19 +698,57 @@ function Neo4jD3(_selector, _options) {
     }
 
     function tick() {
-        tickNodes();
-        tickRelationships();
-    }
+        var nodeTransforms = tickNodes();
 
-    function tickNodes() {
-        if (node) {
-            node.attr('transform', function(d) {
-                return 'translate(' + d.x + ', ' + d.y + ')';
-            });
+        if (!node) {
+            return;
+        }
+
+        var relationshipTransforms = tickRelationships();
+
+        // render all node transforms
+        for (var i = 0; i < node._groups[0].length; i++) {
+            var nodeElement = d3.select(node._groups[0][i]);
+            nodeElement.attr('transform', nodeTransforms[i]);
+        }
+
+        // render all text transforms
+        for (var i = 0; i < relationshipText._groups[0].length; i++) {
+            var text = d3.select(relationshipText._groups[0][i]);
+            text.attr('transform', relationshipTransforms[0][i]);
+        }
+
+        // render all relationship transforms
+        for (var i = 0; i < relationship._groups[0].length; i++) {
+            var outline = d3.select(relationship._groups[0][i]).select('.outline');
+            outline.attr('d', relationshipTransforms[1][i]);
+        }
+
+        // render all overlay transforms
+        for (var i = 0; i < relationshipOverlay._groups[0].length; i++) {
+            var overlay = d3.select(relationshipOverlay._groups[0][i]);
+            overlay.attr('d', relationshipTransforms[2][i]);
         }
     }
 
+    function tickNodes() {
+        if (!node) {
+            return null;
+        }
+
+        var nodeTransforms = [];
+        
+        // process all node transforms
+        node.each(function (d) {
+            nodeTransforms.push('translate(' + d.x + ', ' + d.y + ')');
+        });
+
+        return nodeTransforms;
+    }
+
     function tickRelationships() {
+        var relationshipTransforms = [];
+
         if (relationship) {
             relationship.attr('transform', function(d) {  
                 var center = {x: 0, y: 0},
@@ -737,31 +775,15 @@ function Neo4jD3(_selector, _options) {
             });
 
             const startTime = performance.now();
-            var textTransforms = tickRelationshipsTexts();
-            var outlineTransforms = tickRelationshipsOutlines();
-            var overlayTransforms = tickRelationshipsOverlays();
-
-            // render all text transforms
-            for (var i = 0; i < relationshipText._groups[0].length; i++) {
-                var text = d3.select(relationshipText._groups[0][i]);
-                text.attr('transform', textTransforms[i]);
-            }
-
-            // render all relationship transforms
-            for (var i = 0; i < relationship._groups[0].length; i++) {
-                var outline = d3.select(relationship._groups[0][i]).select('.outline');
-                outline.attr('d', outlineTransforms[i]);
-            }
-
-            // render all overlay transforms
-            for (var i = 0; i < relationshipOverlay._groups[0].length; i++) {
-                var overlay = d3.select(relationshipOverlay._groups[0][i]);
-                overlay.attr('d', overlayTransforms[i]);
-            }
+            relationshipTransforms.push(tickRelationshipsTexts());
+            relationshipTransforms.push(tickRelationshipsOutlines());
+            relationshipTransforms.push(tickRelationshipsOverlays());
 
             const duration = performance.now() - startTime;
             console.log(`tick took ${duration}ms`);
         }
+
+        return relationshipTransforms;
     }
 
     function tickRelationshipsOutlines() {