## Issue fixes
Fixes #162
Fixes #209
Fixes #212
Fixes #213
Fixes #214
Fixes #215

## Benchmarks
### Methodology
Basic methodology is to have very little running on my PC. Then I opened the browser window to a certain size with dev tools open. I run the app in the terminal (and install any npm package changes beforehand) and reload the webpage in my browser. I select all counter filters to show everything. I wait for the graph to settle. I start recording a benchmark in the web browser dev tools. I slowly move a particular operator around in a small circle. I stop recording after a few seconds. I roughly eyeball the performance graph and see roughly on average what the 'ms' improvements are.

### Results
1. Master branch (120ms~ task)
![image](https://user-images.githubusercontent.com/7528322/167265868-2b4c810f-3852-4cb4-9eaa-f1d22a7367bf.png)

2. 'Split processing transforms + rendering into separate steps' (56ms~ task)
![image](https://user-images.githubusercontent.com/7528322/167265963-f62a0f84-e16e-442f-a62d-9a6f550a10e3.png)

3. 'Perform all relationship rendering after all processing' (60ms~ task)
![image](https://user-images.githubusercontent.com/7528322/167266055-ccabf8f5-ab27-4e26-8fc9-81363aca0302.png)

4. 'Perform node rendering after all processing' (70ms~ task)
![image](https://user-images.githubusercontent.com/7528322/167266132-b30b075e-19ca-4a08-ad05-91f0f2633cd3.png)

5. 'High accuracy simulation button' (High accuracy = 65ms~ task, Low accuracy = 60ms~ task)
High
![image](https://user-images.githubusercontent.com/7528322/167266238-a066937f-8330-496b-8138-e979b5b4d553.png)
Low
![image](https://user-images.githubusercontent.com/7528322/167266317-25f114b6-137a-49ee-9181-c558d7ddcc8f.png)

6. 'Split main relationship transform logic' (60ms~ task)
![image](https://user-images.githubusercontent.com/7528322/167266389-f683457a-3a94-4208-a659-a72f4ec44a10.png)

7. 'Use .webp instead of .png' (57ms~ task)
![image](https://user-images.githubusercontent.com/7528322/167266468-d167e9a2-aa39-4806-8ef3-831aea80191a.png)

8. 'Simplify split processing + rendering logic' (57ms~ task)
![image](https://user-images.githubusercontent.com/7528322/167266530-f1b8eeb5-f108-4c13-862b-e7cff9b7eae0.png)

9. 'Update to d3js v5.16.0' (57ms~ task)
![image](https://user-images.githubusercontent.com/7528322/167266596-859507b5-fce8-47ee-8c69-c593fc03b140.png)

10. 'Update d3js to v6.7.0' (57ms~ task)
![image](https://user-images.githubusercontent.com/7528322/167266657-2ae696c4-a0f4-40df-9fa8-346241b83207.png)

11. 'Update d3js to v7.4.4' (57ms~ task)
![image](https://user-images.githubusercontent.com/7528322/167266707-3a31f479-8ea1-4e55-acd4-a105f56e7a09.png)