@BurkusCat 2022-05-07
With all the counters visible, the task time has went from 120ms~ to under 60ms~ (aka 8fps~ to 17fps~).

The performance graph looks completely different now. The script run time is around 7ms~ which is a great sign for trying to get 60fps (there could be micro optimisations here if we ever need to look but script time is very short now relative to the entire frame).

Recalculate style takes 30ms which is a huge amount of time. Layout takes around 10ms. Paint takes around 5ms. My hope is that savings can be made here in the future.